Refactored password reset email sending logic from PasswordResetTokenUtil to AuthenticationService
