+++ src/apps/backend/modules/authentication/internals/password_reset_token/password_reset_token_util.py
@@ -1,6 +1,5 @@
 import hashlib
 import os
-import urllib.parse
 from datetime import datetime, timedelta
 from typing import Any
 
@@ -11,8 +10,6 @@ from modules.authentication.internals.password_reset_token.store.password_reset_
 )
 from modules.authentication.types import PasswordResetToken
 from modules.config.config_service import ConfigService
-from modules.notification.email_service import EmailService
-from modules.notification.types import EmailRecipient, EmailSender, SendEmailParams
 
 
 class PasswordResetTokenUtil:
@@ -54,28 +51,4 @@ class PasswordResetTokenUtil:
             is_expired=PasswordResetTokenUtil.is_token_expired(validated_password_reset_token_data.expires_at),
             expires_at=str(validated_password_reset_token_data.expires_at),
             token=validated_password_reset_token_data.token,
-        )
-
-    @staticmethod
-    def send_password_reset_email(account_id: str, first_name: str, username: str, password_reset_token: str) -> None:
-        web_app_host = ConfigService[str].get_value(key="web_app_host")
-        default_email = ConfigService[str].get_value(key="mailer.default_email")
-        default_email_name = ConfigService[str].get_value(key="mailer.default_email_name")
-        forgot_password_mail_template_id = ConfigService[str].get_value(key="mailer.forgot_password_mail_template_id")
-
-        template_data = {
-            "first_name": first_name,
-            "password_reset_link": f"{web_app_host}/accounts/{account_id}/reset_password?token={urllib.parse.quote(password_reset_token)}",
-            "username": username,
-        }
-
-        password_reset_email_params = SendEmailParams(
-            template_id=forgot_password_mail_template_id,
-            recipient=EmailRecipient(email=username),
-            sender=EmailSender(email=default_email, name=default_email_name),
-            template_data=template_data,
-        )
-
-        EmailService.send_email_for_account(
-            account_id=account_id, bypass_preferences=True, params=password_reset_email_params
         )
